requirements.txt
numpy
pandas
requests
pytest
python-dotenv
openai
beautifulsoup4
google-genai